refactor: Use registrar pattern
